add lab interface
create a new schema lab_interface
add all tables for the setup_conf_idx to the interface module
add all the configuration tables to the interface module
