+++ Experiments/Calibrate.py
@@ -387,7 +387,7 @@ class Experiment:
         self.logger.put(
             table="PortCalibration",
             tuple=key,
-            schema="behavior",
+            schema="interface",
             priority=5,
             ignore_extra_fields=True,
             validate=True,
@@ -396,7 +396,7 @@ class Experiment:
         )
         self.logger.put(
             table="PortCalibration.Liquid",
-            schema="behavior",
+            schema="interface",
             replace=True,
             ignore_extra_fields=True,
             tuple=dict(
